requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
openai==1.35.15
openai-whisper==20231117
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
setuptools==72.1.0
sniffio==1.3.1
sounddevice==0.4.7
sympy==1.13.1
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wavio==0.0.9
